backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.115.13
google-auth==2.40.3
google-genai==1.21.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.0
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
starlette==0.46.2
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
websockets==15.0.1
